red_team_tools.py
import requests
from utils.network_utils import analyze_traffic, detect_intrusion, monitor_network
from utils.forensic_utils import forensic_data_schema
from utils.crypto_utils import encrypt_data, decrypt_data
from utils.reverse_engineering_utils import disassemble_binary, analyze_executable
from utils.logging_utils import log_action, log_error, log_info
from dynamic_code_exec import execute_dynamic_code
from polymorphic_code_gen import generate_polymorphic_code
from api_interaction import send_api_request, receive_api_response
from data.payloads import payload_schema
from config import api_key

class RedTeamTools:
    def __init__(self):
        self.api_key = api_key

    def conduct_recon(self, target):
        log_info(f"Conducting reconnaissance on {target}")
        # Implement reconnaissance logic here
        pass

    def gain_access(self, target):
        log_info(f"Gaining access to {target}")
        # Implement access logic here, possibly using exploits or credentials
        pass

    def escalate_privileges(self, target):
        log_info(f"Escalating privileges on {target}")
        # Implement privilege escalation logic here
        pass

    def maintain_access(self, target):
        log_info(f"Maintaining access to {target}")
        # Implement logic to maintain access, such as creating backdoors
        pass

    def cover_tracks(self, target):
        log_info(f"Covering tracks on {target}")
        # Implement logic to erase logs and other forensic evidence
        pass

    def simulate_attack(self, scenario):
        log_info(f"Simulating attack scenario: {scenario}")
        # Implement attack simulation logic here
        pass

    def execute_payload(self, target, payload):
        log_info(f"Executing payload on {target}")
        # Implement payload execution logic here
        pass

    def dynamic_execution(self, code_snippet):
        log_info("Performing dynamic code execution")
        # Execute dynamic code using the exec() function
        execute_dynamic_code(code_snippet)

    def generate_polymorphic_code(self):
        log_info("Generating polymorphic code")
        # Generate polymorphic code to evade detection
        return generate_polymorphic_code()

    def interact_with_api(self, data):
        log_info("Interacting with external API")
        # Send and receive data from an external API
        response = send_api_request(self.api_key, data)
        return receive_api_response(response)

    def post_mortem_analysis(self, scenario, outcome):
        log_info(f"Conducting post-mortem analysis for scenario: {scenario}")
        # Analyze the effectiveness and outcomes of the scenario
        pass

    def monitor_for_attacks(self):
        log_info("Monitoring for incoming attacks")
        # Implement monitoring logic here
        pass

    def search_for_vulnerabilities(self, target):
        log_info(f"Searching for vulnerabilities in {target}")
        # Implement vulnerability search logic here
        pass

# Example usage of the RedTeamTools class
if __name__ == "__main__":
    red_team = RedTeamTools()
    target_system = "http://example.com"
    red_team.conduct_recon(target_system)
    red_team.gain_access(target_system)
    red_team.escalate_privileges(target_system)
    red_team.maintain_access(target_system)
    red_team.cover_tracks(target_system)
    red_team.simulate_attack("CTF Scenario")
    red_team.execute_payload(target_system, payload_schema)
    dynamic_code = "print('This is a dynamically executed print statement')"
    red_team.dynamic_execution(dynamic_code)
    polymorphic_code = red_team.generate_polymorphic_code()
    api_data = {"data": "example"}
    api_response = red_team.interact_with_api(api_data)
    red_team.post_mortem_analysis("CTF Scenario", "Success")
    red_team.monitor_for_attacks()
    red_team.search_for_vulnerabilities(target_system)