tests/test_red_team_tools.py
import unittest
from red_team_tools import conduct_recon, gain_access, escalate_privileges
from utils.helpers import api_key, log_action
from utils.network_utils import analyze_traffic
from utils.forensic_utils import analyze_logs
from data.configurations import config_schema

class TestRedTeamTools(unittest.TestCase):

    def setUp(self):
        # Setup with necessary configurations and API key
        self.api_key = api_key
        self.config = config_schema

    def test_conduct_recon(self):
        # Test the reconnaissance function for gathering intelligence
        target = 'test_target'
        recon_data = conduct_recon(target, self.api_key)
        self.assertIsNotNone(recon_data, "Reconnaissance data should not be None")
        log_action("Conduct Recon Test Passed", recon_data)

    def test_gain_access(self):
        # Test the function for gaining access to a system
        target = 'test_target'
        access_data = gain_access(target, self.api_key)
        self.assertTrue(access_data['success'], "Access should be successful")
        log_action("Gain Access Test Passed", access_data)

    def test_escalate_privileges(self):
        # Test the privilege escalation function
        target = 'test_target'
        privileges_data = escalate_privileges(target, self.api_key)
        self.assertTrue(privileges_data['success'], "Privilege escalation should be successful")
        log_action("Escalate Privileges Test Passed", privileges_data)

    def test_analyze_traffic(self):
        # Test network traffic analysis
        traffic_data = analyze_traffic(self.config)
        self.assertIn('traffic_analysis', traffic_data, "Traffic analysis data should be present")
        log_action("Analyze Traffic Test Passed", traffic_data)

    def test_analyze_logs(self):
        # Test log analysis for incident response
        logs_data = analyze_logs(self.config)
        self.assertIn('incident_detected', logs_data, "Incident detection data should be present")
        log_action("Analyze Logs Test Passed", logs_data)

if __name__ == '__main__':
    unittest.main()